Add capability to continue when no versions found for a subfolder
When trying to recover a whole bucket objects and it fails to find Versions in a page, the command exits and fails. This is a bit noisy because you cannot move forward to other folders/files. Just let to have the choice to "continue", instead of automatically stopping the program, will provide the capability to move on the following files in the bucket.